Make nftables metrics linux-specific
