+++ pkg/proxy/metrics/metrics_other.go
@@ -1,0 +1,29 @@
+//go:build !linux
+
+/*
+Copyright The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package metrics
+
+import "k8s.io/component-base/metrics"
+
+func newNFAcctMetricCollector(counter string, description *metrics.Desc) metrics.StableCollector {
+	return nil
+}
+
+func newNFTablesCounterMetricCollector(counters map[string]string, description *metrics.Desc) metrics.StableCollector {
+	return nil
+}
